[Bugfix] Fix compilation error since python 3.11 and macOS
bfloat16 could not be built since Python 3.11, for Py_TYPE has been
deprecated since Python 3.9, see PEP 674 and Common Object
Structures docs for details. This issue was covered in #5.
bfloat16 could not be built on macOS, because C++11 flag is not
enabled by default in clang. This issue was covered in #7.